--- server/attractions/src/main/java/com/example/codeclan/attractions/repositories/AccessibilityRepository.java
@@ -1,9 +1,0 @@
-package com.example.codeclan.attractions.repositories;
-
-import com.example.codeclan.attractions.models.Accessibility;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AccessibilityRepository extends JpaRepository <Accessibility, Long> {
-}
