+++ server/attractions/src/main/resources/application.properties
@@ -2,6 +2,6 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/attractions
 spring.datasource.username=postgres
 spring.datasource.password=postgres
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL9Dialect
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
